requirements.txt
alabaster==0.7.12
altgraph==0.17
appdirs==1.4.3
arrow==0.15.6
asgiref==3.2.7
astroid==2.4.1
atomicwrites==1.4.0
attrs==19.3.0
Automat==20.2.0
autopep8==1.5.3
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.0
bleach==3.1.5
Brotli==1.0.7
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
colorama==0.4.3
commonmark==0.9.1
constantly==15.1.0
cryptography==2.9.2
cycler==0.10.0
dash==1.12.0
dash-core-components==1.10.0
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
Django==3.0.7
dj-database-url==0.5.0
django-plotly-dash>=1.3.1
docutils==0.16
entrypoints==0.3
filelock==3.0.12
flake8==3.8.2
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
gunicorn==20.0.4
helpdev==0.7.1
hyperlink==19.0.0
idna==2.9
imagesize==1.2.0
incremental==17.5.0
inflection==0.4.0
intervaltree==3.0.2
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.11.2
jsonschema==3.2.0
keyring==21.2.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
multidict==4.7.5
nbconvert==5.6.1
nbformat==5.0.6
numpy==1.18.3
numpydoc==1.0.0
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.2
pathtools==0.1.2
pefile==2019.4.18
pexpect==4.8.0
pickleshare==0.7.5
plotly==4.8.0
pluggy==0.13.1
pprintpp==0.4.0
prompt-toolkit==3.0.5
psutil==5.7.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.5.1
pydocstyle==5.0.2
pyflakes>=2.1.1
Pygments==2.6.1
PyHamcrest==2.0.2
PyInstaller==3.6
pylint==2.5.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
pytz==2020.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole==4.7.4
QtPy==1.9.0
requests==2.23.0
retrying==1.3.3
rich==0.8.13
rope==0.17.0
scipy==1.4.1
six==1.14.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
soupsieve==2.0
Sphinx==3.0.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
statistics==1.0.3.5
stravalib==0.10.2
stripe==2.48.0
swagger-codegen==0.1.3
testpath==0.4.4
toml==0.10.0
tornado==6.0.4
traitlets==4.3.3
Twisted==20.3.0
typer==0.1.1
typing-extensions==3.7.4.2
units==0.7
urllib3==1.25.9
virtualenv==20.0.18
watchdog==0.10.2
wcwidth==0.2.2
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.1.0
wrapt==1.12.1
yapf==0.30.0
zope.interface==5.1.0
autobahn==20.6.1
channels==2.4.0
daphne==2.5.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyopenssl==19.1.0
service-identity==18.1.0
txaio==20.4.1
aioredis==1.3.1
async-timeout==3.0.1
channels-redis==2.4.2
hiredis==1.0.1
msgpack==0.6.2
django-tables2==2.3.1
django-crispy-forms==1.9.1
dash-bootstrap-components==0.10.2
django-heroku==0.3.1
django_compressor==2.4
django-appconf==1.0.3
rcssmin==1.0.6
rjsmin==1.1.0
python-decouple==3.3
asyncio==3.4.3
aiohttp==3.6.2
yarl==1.4.2
django-allauth==0.42.0
oauthlib==3.1.0
python3-openid==3.1.0
requests-oauthlib==1.3.0
polyline==1.4.0
pgcopy==1.4.1
django-social-share==1.4.0